requirements.txt
pyyaml
importlib
click
requests
hjson
N2G
python-igraph